refactor: initialize map on mobile devices when opened
