+++ src/layouts/shared/menu-utils.tsx
@@ -1,6 +1,7 @@
-import {and} from '@atb/utils/css';
-import {MutableRefObject, useCallback, useEffect, useState} from 'react';
+import { and } from '@atb/utils/css';
+import { MutableRefObject, useCallback, useEffect, useState } from 'react';
 import style from './page-header.module.css';
+import useMediaQuery from '@atb/utils/user-media-query';
 
 export function Hamburger({
   isOpen,
@@ -128,48 +129,4 @@ export function useTogglableBurgerMenu<
     setMenuVisible,
     tabIndex,
   };
-}
-
-function hasWindow() {
-  return typeof window !== 'undefined';
-}
-
-const useMediaQuery = (mediaQuery: string) => {
-  const [isMatched, setMatched] = useState(() => {
-    if (!hasWindow()) return false;
-    return Boolean(window.matchMedia(mediaQuery).matches);
-  });
-
-  useEffect(() => {
-    if (!hasWindow()) return;
-    const mediaQueryList = window.matchMedia(mediaQuery);
-    const documentChangeHandler = () =>
-      setMatched(Boolean(mediaQueryList.matches));
-    listenTo(mediaQueryList, documentChangeHandler);
-
-    documentChangeHandler();
-    return () => removeListener(mediaQueryList, documentChangeHandler);
-  }, [mediaQuery]);
-
-  return isMatched;
-};
-
-function listenTo(
-  matcher: MediaQueryList,
-  cb: (ev: MediaQueryListEvent) => void,
-) {
-  if ('addEventListener' in (matcher as any)) {
-    return matcher.addEventListener('change', cb);
-  }
-  return matcher.addListener(cb);
-}
-
-function removeListener(
-  matcher: MediaQueryList,
-  cb: (ev: MediaQueryListEvent) => void,
-) {
-  if ('removeEventListener' in (matcher as any)) {
-    return matcher.removeEventListener('change', cb);
-  }
-  return matcher.removeListener(cb);
 }
