+++ src/utils/user-media-query.ts
@@ -1,0 +1,45 @@
+import { useEffect, useState } from 'react';
+
+export default function useMediaQuery(mediaQuery: string) {
+  const [isMatched, setMatched] = useState(() => {
+    if (!hasWindow()) return false;
+    return Boolean(window.matchMedia(mediaQuery).matches);
+  });
+
+  useEffect(() => {
+    if (!hasWindow()) return;
+    const mediaQueryList = window.matchMedia(mediaQuery);
+    const documentChangeHandler = () =>
+      setMatched(Boolean(mediaQueryList.matches));
+    listenTo(mediaQueryList, documentChangeHandler);
+
+    documentChangeHandler();
+    return () => removeListener(mediaQueryList, documentChangeHandler);
+  }, [mediaQuery]);
+
+  return isMatched;
+}
+
+function hasWindow() {
+  return typeof window !== 'undefined';
+}
+
+function listenTo(
+  matcher: MediaQueryList,
+  cb: (ev: MediaQueryListEvent) => void,
+) {
+  if ('addEventListener' in (matcher as any)) {
+    return matcher.addEventListener('change', cb);
+  }
+  return matcher.addListener(cb);
+}
+
+function removeListener(
+  matcher: MediaQueryList,
+  cb: (ev: MediaQueryListEvent) => void,
+) {
+  if ('removeEventListener' in (matcher as any)) {
+    return matcher.removeEventListener('change', cb);
+  }
+  return matcher.removeListener(cb);
+}
